ready to pull

+++ src/main/java/org/example/final1/model/BookDto.java
@@ -18,7 +18,7 @@ public class BookDto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "book_id")
-    private int book_Id;
+    private Long book_Id;
 
     @Column(name = "book_image", nullable = false, length = 255)
     private String book_image;
