+++ src/main/java/org/example/final1/model/CategoryDto.java
@@ -18,7 +18,7 @@ public class CategoryDto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private int category_id;
+    private Long category_id;
 
     @Column(name = "category_name", nullable = false, length = 50)
     private String category_name;
